refactor: remove AtomicBroadcast and FedimintConsensus, merge functionality into ConsensusServer
