+++ fedimint-server/src/atomic_broadcast/network.rs
@@ -4,7 +4,6 @@ use bitcoin_hashes_12::{sha256, Hash};
 use fedimint_core::net::peers::IPeerConnections;
 use parity_scale_codec::{Decode, Encode, IoReader};
 
-use super::conversion::to_peer_id;
 use super::data_provider::UnitData;
 use super::keychain::Keychain;
 use super::{Message, Recipient};
@@ -48,7 +47,9 @@ impl aleph_bft::Network<NetworkData> for Network {
     fn send(&self, network_data: NetworkData, recipient: aleph_bft::Recipient) {
         // convert from aleph_bft::Recipient to session::Recipient
         let recipient = match recipient {
-            aleph_bft::Recipient::Node(node_index) => Recipient::Peer(to_peer_id(node_index)),
+            aleph_bft::Recipient::Node(node_index) => {
+                Recipient::Peer(super::to_peer_id(node_index))
+            }
             aleph_bft::Recipient::Everyone => Recipient::Everyone,
         };
