+++ fedimint-server/src/consensus/server.rs
@@ -2,30 +2,48 @@ use std::collections::{BTreeMap, HashMap};
 use std::sync::Arc;
 use std::time::Duration;
 
-use anyhow::bail;
-use async_channel::Sender;
+use aleph_bft::Keychain as KeychainTrait;
+use anyhow::{anyhow, bail};
+use async_channel::{Receiver, Sender};
 use bitcoin_hashes::sha256;
-use fedimint_core::api::{GlobalFederationApi, WsFederationApi};
+use fedimint_core::api::{FederationApiExt, GlobalFederationApi, WsFederationApi};
+use fedimint_core::block::{AcceptedItem, Block, SchnorrSignature, SignedBlock};
 use fedimint_core::config::ServerModuleInitRegistry;
-use fedimint_core::db::{apply_migrations, Database};
-use fedimint_core::epoch::{ConsensusItem, SerdeSignatureShare};
+use fedimint_core::db::{apply_migrations, Database, DatabaseTransaction};
+use fedimint_core::encoding::Decodable;
+use fedimint_core::endpoint_constants::AWAIT_SIGNED_BLOCK_ENDPOINT;
+use fedimint_core::epoch::{ConsensusItem, SerdeSignature, SerdeSignatureShare};
 use fedimint_core::fmt_utils::OptStacktrace;
-use fedimint_core::module::registry::{ModuleRegistry, ServerModuleRegistry};
-use fedimint_core::task::{sleep, RwLock, TaskGroup, TaskHandle};
+use fedimint_core::module::audit::Audit;
+use fedimint_core::module::registry::{
+    ModuleDecoderRegistry, ModuleRegistry, ServerModuleRegistry,
+};
+use fedimint_core::module::{ApiRequestErased, SerdeModuleEncoding};
+use fedimint_core::query::FilterMap;
+use fedimint_core::task::{sleep, spawn, RwLock, TaskGroup, TaskHandle};
+use fedimint_core::util::SafeUrl;
 use fedimint_core::{timing, PeerId};
-use tracing::{info, warn};
+use futures::StreamExt;
+use tokio::sync::watch;
+use tracing::{debug, info, warn};
 
-use crate::atomic_broadcast::{AtomicBroadcast, Keychain, Message};
+use crate::atomic_broadcast::data_provider::{DataProvider, UnitData};
+use crate::atomic_broadcast::finalization_handler::FinalizationHandler;
+use crate::atomic_broadcast::network::Network;
+use crate::atomic_broadcast::spawner::Spawner;
+use crate::atomic_broadcast::{to_node_index, Keychain, Message};
 use crate::config::ServerConfig;
-use crate::consensus::FedimintConsensus;
+use crate::consensus::process_transaction_with_dbtx;
 use crate::db::{
-    get_global_database_migrations, ClientConfigSignatureKey, GLOBAL_DATABASE_VERSION,
+    get_global_database_migrations, AcceptedItemKey, AcceptedItemPrefix, AcceptedTransactionKey,
+    AlephUnitsPrefix, ClientConfigSignatureKey, ClientConfigSignatureShareKey,
+    ClientConfigSignatureSharePrefix, SignedBlockKey, SignedBlockPrefix, GLOBAL_DATABASE_VERSION,
 };
 use crate::fedimint_core::encoding::Encodable;
 use crate::net::api::{ConsensusApi, ExpiringCache, InvitationCodesTracker};
 use crate::net::connect::{Connector, TlsTcpConnector};
 use crate::net::peers::{DelayCalculator, PeerConnector, ReconnectPeerConnections};
-use crate::{LOG_CONSENSUS, LOG_CORE};
+use crate::{atomic_broadcast, LOG_CONSENSUS, LOG_CORE};
 
 /// How many txs can be stored in memory before blocking the API
 const TRANSACTION_BUFFER: usize = 1000;
@@ -34,16 +52,15 @@ pub(crate) type LatestContributionByPeer = HashMap<PeerId, u64>;
 
 /// Runs the main server consensus loop
 pub struct ConsensusServer {
-    /// Allows clients to access consensus state
-    pub db: Database,
-    pub modules: ServerModuleRegistry,
-    pub consensus_api: ConsensusApi,
-    /// Aleph BFT instance
-    pub atomic_broadcast: AtomicBroadcast,
-    /// Our configuration
-    pub cfg: ServerConfig,
-    /// tracks the last session a message was received by peer
-    pub latest_contribution_by_peer: Arc<RwLock<LatestContributionByPeer>>,
+    modules: ServerModuleRegistry,
+    db: Database,
+    connections: ReconnectPeerConnections<Message>,
+    keychain: Keychain,
+    client_cfg_hash: sha256::Hash,
+    api_endpoints: Vec<(PeerId, SafeUrl)>,
+    cfg: ServerConfig,
+    submission_receiver: Receiver<ConsensusItem>,
+    latest_contribution_by_peer: Arc<RwLock<LatestContributionByPeer>>,
 }
 
 impl ConsensusServer {
@@ -53,7 +70,7 @@ impl ConsensusServer {
         db: Database,
         module_inits: ServerModuleInitRegistry,
         task_group: &mut TaskGroup,
-    ) -> anyhow::Result<Self> {
+    ) -> anyhow::Result<(Self, ConsensusApi)> {
         let connector: PeerConnector<Message> =
             TlsTcpConnector::new(cfg.tls_config(), cfg.local.identity).into_dyn();
 
@@ -78,7 +95,7 @@ impl ConsensusServer {
         connector: PeerConnector<Message>,
         delay_calculator: DelayCalculator,
         task_group: &mut TaskGroup,
-    ) -> anyhow::Result<Self> {
+    ) -> anyhow::Result<(Self, ConsensusApi)> {
         // We need four peers to run the atomic broadcast
         assert!(cfg.consensus.api_endpoints.len() >= 4);
 
@@ -145,13 +162,6 @@ impl ConsensusServer {
         )
         .await;
 
-        let atomic_broadcast = AtomicBroadcast::new(
-            keychain,
-            db.clone(),
-            connections.clone(),
-            submission_receiver,
-        );
-
         // Build API that can handle requests
         let latest_contribution_by_peer = Default::default();
 
@@ -177,23 +187,11 @@ impl ConsensusServer {
             modules.clone(),
             cfg.clone(),
             consensus_api.client_cfg.consensus_hash(),
-            submission_sender,
+            submission_sender.clone(),
         )
         .await;
 
-        Ok(ConsensusServer {
-            atomic_broadcast,
-            db,
-            consensus_api,
-            cfg: cfg.clone(),
-            latest_contribution_by_peer,
-            modules,
-        })
-    }
-
-    pub async fn run_consensus(&self, task_handle: TaskHandle) -> anyhow::Result<()> {
-        let api_endpoints: Vec<_> = self
-            .cfg
+        let api_endpoints: Vec<_> = cfg
             .consensus
             .api_endpoints
             .clone()
@@ -201,25 +199,35 @@ impl ConsensusServer {
             .map(|(id, node)| (id, node.url))
             .collect();
 
-        let federation_api = WsFederationApi::new(api_endpoints.clone());
+        let consensus_server = ConsensusServer {
+            connections,
+            db,
+            keychain,
+            client_cfg_hash: consensus_api.client_cfg.consensus_hash(),
+            api_endpoints,
+            cfg: cfg.clone(),
+            submission_receiver,
+            latest_contribution_by_peer,
+            modules,
+        };
 
-        confirm_consensus_config_hash(&federation_api, self.cfg.consensus.consensus_hash()).await?;
+        Ok((consensus_server, consensus_api))
+    }
 
-        // TODO: latest contribution by peer
+    pub async fn run_consensus(&self, task_handle: TaskHandle) -> anyhow::Result<()> {
+        self.confirm_consensus_config_hash().await?;
+
         while !task_handle.is_shutting_down() {
-            let consensus = FedimintConsensus::load_current_session(
-                self.cfg.clone(),
-                self.modules.clone(),
-                self.db.clone(),
-                self.consensus_api.client_cfg.consensus_hash(),
-                self.latest_contribution_by_peer.clone(),
-            );
+            let session_index = self
+                .db
+                .begin_transaction()
+                .await
+                .find_by_prefix(&SignedBlockPrefix)
+                .await
+                .count()
+                .await as u64;
 
-            let federation_api = WsFederationApi::new(api_endpoints.clone());
-
-            self.atomic_broadcast
-                .run_session(consensus.await, federation_api)
-                .await?;
+            self.run_session(session_index).await?;
 
             info!(target: LOG_CONSENSUS, "Session completed");
         }
@@ -228,31 +236,454 @@ impl ConsensusServer {
 
         Ok(())
     }
-}
 
-async fn confirm_consensus_config_hash(
-    api: &WsFederationApi,
-    our_hash: sha256::Hash,
-) -> anyhow::Result<()> {
-    info!(target: LOG_CONSENSUS, "Waiting for peers config {our_hash}");
+    async fn confirm_consensus_config_hash(&self) -> anyhow::Result<()> {
+        let our_hash = self.cfg.consensus.consensus_hash();
+        let federation_api = WsFederationApi::new(self.api_endpoints.clone());
 
-    loop {
-        match api.consensus_config_hash().await {
-            Ok(consensus_hash) => {
-                if consensus_hash != our_hash {
-                    bail!("Our consensus config doesn't match peers!")
+        info!(target: LOG_CONSENSUS, "Waiting for peers config {our_hash}");
+
+        loop {
+            match federation_api.consensus_config_hash().await {
+                Ok(consensus_hash) => {
+                    if consensus_hash != our_hash {
+                        bail!("Our consensus config doesn't match peers!")
+                    }
+
+                    info!(target: LOG_CONSENSUS, "Confirmed peers config {our_hash}");
+
+                    return Ok(());
                 }
-
-                info!(target: LOG_CONSENSUS, "Confirmed peers config {our_hash}");
-
-                return Ok(());
+                Err(e) => {
+                    warn!(target: LOG_CONSENSUS, "Could not check consensus config hash: {}", OptStacktrace(e))
+                }
             }
-            Err(e) => {
-                warn!(target: LOG_CONSENSUS, "Could not check consensus config hash: {}", OptStacktrace(e))
+
+            sleep(Duration::from_millis(100)).await;
+        }
+    }
+
+    pub async fn run_session(&self, session_index: u64) -> anyhow::Result<()> {
+        // if all nodes are correct the session will take 45 to 60 seconds. The
+        // more nodes go offline the longer the session will take to complete.
+        const EXPECTED_ROUNDS_PER_SESSION: usize = 45 * 4;
+        // this constant needs to be 3000 or less to guarantee that the session
+        // can never reach MAX_ROUNDs.
+        const EXPONENTIAL_SLOWDOWN_OFFSET: usize = 3 * EXPECTED_ROUNDS_PER_SESSION;
+        const MAX_ROUND: u16 = 5000;
+        const ROUND_DELAY: f64 = 250.0;
+        const BASE: f64 = 1.01;
+
+        // this is the minimum number of unit data that will be ordered before we reach
+        // the EXPONENTIAL_SLOWDOWN_OFFSET even if f peers do not attach unit data
+        let batches_per_session = EXPECTED_ROUNDS_PER_SESSION * self.keychain.peer_count();
+
+        // In order to bound a sessions RAM consumption we need to bound its number of
+        // units and therefore its number of rounds. Since we use a session to
+        // create a threshold signature for the corresponding block we have to
+        // guarantee that an attacker cannot exhaust our memory by preventing the
+        // creation of a threshold signature, thereby keeping the session open
+        // indefinitely. Hence we increase the delay between rounds exponentially
+        // such that MAX_ROUND would only be reached after roughly 350 years.
+        // In case of such an attack the broadcast stops ordering any items until the
+        // attack subsides as not items are ordered while the signatures are collected.
+        let mut delay_config = aleph_bft::default_delay_config();
+        delay_config.unit_creation_delay = std::sync::Arc::new(|round_index| {
+            let delay = if round_index == 0 {
+                0.0
+            } else {
+                ROUND_DELAY
+                    * BASE.powf(round_index.saturating_sub(EXPONENTIAL_SLOWDOWN_OFFSET) as f64)
+            };
+
+            Duration::from_millis(delay.round() as u64)
+        });
+
+        let config = aleph_bft::create_config(
+            self.keychain.peer_count().into(),
+            self.keychain.peer_id().to_usize().into(),
+            session_index,
+            MAX_ROUND,
+            delay_config,
+            Duration::from_secs(100 * 365 * 24 * 60 * 60),
+        )
+        .expect("Config is valid");
+
+        // the number of units ordered in a single aleph session is bounded
+        let (unit_data_sender, unit_data_receiver) = async_channel::unbounded();
+        let (signature_sender, signature_receiver) = watch::channel(None);
+        let (terminator_sender, terminator_receiver) = futures::channel::oneshot::channel();
+
+        let (loader, saver) = atomic_broadcast::backup::load_session(self.db.clone()).await;
+
+        let aleph_handle = spawn(
+            "aleph run session",
+            aleph_bft::run_session(
+                config,
+                aleph_bft::LocalIO::new(
+                    DataProvider::new(self.submission_receiver.clone(), signature_receiver),
+                    FinalizationHandler::new(unit_data_sender),
+                    saver,
+                    loader,
+                ),
+                Network::new(self.connections.clone()),
+                self.keychain.clone(),
+                Spawner::new(),
+                aleph_bft_types::Terminator::create_root(terminator_receiver, "Terminator"),
+            ),
+        )
+        .expect("some handle on non-wasm");
+
+        let signed_block = self
+            .complete_signed_block(
+                session_index,
+                batches_per_session,
+                unit_data_receiver,
+                signature_sender,
+            )
+            .await?;
+
+        terminator_sender.send(()).ok();
+        aleph_handle.await.ok();
+
+        // Only call this after aleph bft has shutdown to avoid write-write conflicts
+        // for the aleph bft units
+        self.complete_session(session_index, signed_block).await;
+
+        Ok(())
+    }
+
+    pub async fn complete_signed_block(
+        &self,
+        session_index: u64,
+        batches_per_block: usize,
+        unit_data_receiver: Receiver<(UnitData, PeerId)>,
+        signature_sender: watch::Sender<Option<SchnorrSignature>>,
+    ) -> anyhow::Result<SignedBlock> {
+        let mut num_batches = 0;
+        let mut item_index = 0;
+
+        // we build a block out of the ordered batches until either we have processed
+        // n_batches_per_block blocks or a signed block arrives from our peers
+        while num_batches < batches_per_block {
+            tokio::select! {
+                unit_data = unit_data_receiver.recv() => {
+                    if let (UnitData::Batch(bytes), peer) = unit_data? {
+                        if let Ok(items) = Vec::<ConsensusItem>::consensus_decode(&mut bytes.as_slice(), &self.decoders()){
+                            for item in items {
+                                if self.process_consensus_item(
+                                    session_index,
+                                    item_index,
+                                    item.clone(),
+                                    peer
+                                ).await
+                                .is_ok() {
+                                    item_index += 1;
+                                }
+                            }
+                        }
+                        num_batches += 1;
+                    }
+                },
+                signed_block = self.request_signed_block(session_index) => {
+                    let partial_block = self.build_block().await.items;
+
+                    let (processed, unprocessed) = signed_block.block.items.split_at(partial_block.len());
+
+                    assert!(processed.iter().eq(partial_block.iter()));
+
+                    for accepted_item in unprocessed {
+                        let result = self.process_consensus_item(
+                            session_index,
+                            item_index,
+                            accepted_item.item.clone(),
+                            accepted_item.peer
+                        ).await;
+
+                        assert!(result.is_ok());
+
+                        item_index += 1;
+                    }
+
+                    return Ok(signed_block);
+                }
             }
         }
 
-        sleep(Duration::from_millis(100)).await;
+        let block = self.build_block().await;
+        let header = block.header(session_index);
+
+        // we send our own signature to the data provider to be broadcasted
+        signature_sender.send(Some(self.keychain.sign(&header)))?;
+
+        let mut signatures = BTreeMap::new();
+
+        // we collect the ordered signatures until we either obtain a threshold
+        // signature or a signed block arrives from our peers
+        while signatures.len() < self.keychain.threshold() {
+            tokio::select! {
+                unit_data = unit_data_receiver.recv() => {
+                    if let (UnitData::Signature(signature), peer) = unit_data? {
+                        if self.keychain.verify(&header, &signature, to_node_index(peer)){
+                            // since the signature is valid the node index can be converted to a peer id
+                            signatures.insert(peer, signature);
+                        }
+                    }
+                }
+                signed_block = self.request_signed_block(session_index) => {
+                    // We check that the block we have created agrees with the federations consensus
+                    assert!(header == signed_block.block.header(session_index));
+
+                    return Ok(signed_block);
+                }
+            }
+        }
+
+        Ok(SignedBlock { block, signatures })
+    }
+
+    fn decoders(&self) -> ModuleDecoderRegistry {
+        self.modules.decoder_registry()
+    }
+
+    pub async fn build_block(&self) -> Block {
+        let items = self
+            .db
+            .begin_transaction()
+            .await
+            .find_by_prefix(&AcceptedItemPrefix)
+            .await
+            .map(|entry| entry.1)
+            .collect()
+            .await;
+
+        Block { items }
+    }
+
+    pub async fn complete_session(&self, session_index: u64, signed_block: SignedBlock) {
+        let mut dbtx = self.db.begin_transaction().await;
+
+        dbtx.remove_by_prefix(&AlephUnitsPrefix).await;
+
+        dbtx.remove_by_prefix(&AcceptedItemPrefix).await;
+
+        if dbtx
+            .insert_entry(&SignedBlockKey(session_index), &signed_block)
+            .await
+            .is_some()
+        {
+            panic!("We tried to overwrite a signed block");
+        }
+
+        dbtx.commit_tx_result()
+            .await
+            .expect("This is the only place where we write to this key");
+    }
+
+    pub async fn process_consensus_item(
+        &self,
+        session_index: u64,
+        item_index: u64,
+        item: ConsensusItem,
+        peer: PeerId,
+    ) -> anyhow::Result<()> {
+        let _timing /* logs on drop */ = timing::TimeReporter::new("process_consensus_item");
+
+        debug!("Peer {peer}: {}", super::debug::item_message(&item));
+
+        self.latest_contribution_by_peer
+            .write()
+            .await
+            .insert(peer, session_index);
+
+        let mut dbtx = self.db.begin_transaction().await;
+
+        if let Some(accepted_item) = dbtx
+            .get_value(&AcceptedItemKey(item_index.to_owned()))
+            .await
+        {
+            if accepted_item.item == item && accepted_item.peer == peer {
+                return Ok(());
+            }
+
+            bail!("Consensus item was discarded before recovery");
+        }
+
+        self.process_consensus_item_with_db_transaction(&mut dbtx, item.clone(), peer)
+            .await?;
+
+        dbtx.insert_entry(&AcceptedItemKey(item_index), &AcceptedItem { item, peer })
+            .await;
+
+        let mut audit = Audit::default();
+
+        for (module_instance_id, _, module) in self.modules.iter_modules() {
+            module
+                .audit(
+                    &mut dbtx.with_module_prefix(module_instance_id),
+                    &mut audit,
+                    module_instance_id,
+                )
+                .await
+        }
+
+        if audit.net_assets().milli_sat < 0 {
+            panic!("Balance sheet of the fed has gone negative, this should never happen! {audit}")
+        }
+
+        dbtx.commit_tx_result()
+            .await
+            .expect("Committing consensus epoch failed");
+
+        Ok(())
+    }
+
+    async fn process_consensus_item_with_db_transaction(
+        &self,
+        dbtx: &mut DatabaseTransaction<'_>,
+        consensus_item: ConsensusItem,
+        peer_id: PeerId,
+    ) -> anyhow::Result<()> {
+        // We rely on decoding rejecting any unknown module instance ids to avoid
+        // peer-triggered panic here
+        self.decoders().assert_reject_mode();
+
+        match consensus_item {
+            ConsensusItem::Module(module_item) => {
+                let moduletx = &mut dbtx.with_module_prefix(module_item.module_instance_id());
+
+                self.modules
+                    .get_expect(module_item.module_instance_id())
+                    .process_consensus_item(moduletx, module_item, peer_id)
+                    .await
+            }
+            ConsensusItem::Transaction(transaction) => {
+                if dbtx
+                    .get_value(&AcceptedTransactionKey(transaction.tx_hash()))
+                    .await
+                    .is_some()
+                {
+                    bail!("The transaction is already accepted");
+                }
+
+                let txid = transaction.tx_hash();
+                let modules_ids = transaction
+                    .outputs
+                    .iter()
+                    .map(|output| output.module_instance_id())
+                    .collect::<Vec<_>>();
+
+                process_transaction_with_dbtx(self.modules.clone(), dbtx, transaction).await?;
+
+                dbtx.insert_entry(&AcceptedTransactionKey(txid), &modules_ids)
+                    .await;
+
+                Ok(())
+            }
+            ConsensusItem::ClientConfigSignatureShare(signature_share) => {
+                if dbtx
+                    .get_isolated()
+                    .get_value(&ClientConfigSignatureKey)
+                    .await
+                    .is_some()
+                {
+                    bail!("Client config is already signed");
+                }
+
+                if dbtx
+                    .get_value(&ClientConfigSignatureShareKey(peer_id))
+                    .await
+                    .is_some()
+                {
+                    bail!("Already received a valid signature share for this peer");
+                }
+
+                let pks = self.cfg.consensus.auth_pk_set.clone();
+
+                if !pks
+                    .public_key_share(peer_id.to_usize())
+                    .verify(&signature_share.0, self.client_cfg_hash)
+                {
+                    bail!("Client config signature share is invalid");
+                }
+
+                // we have received the first valid signature share for this peer
+                dbtx.insert_new_entry(&ClientConfigSignatureShareKey(peer_id), &signature_share)
+                    .await;
+
+                // collect all valid signature shares received previously
+                let signature_shares = dbtx
+                    .find_by_prefix(&ClientConfigSignatureSharePrefix)
+                    .await
+                    .map(|(key, share)| (key.0.to_usize(), share.0))
+                    .collect::<Vec<_>>()
+                    .await;
+
+                if signature_shares.len() <= pks.threshold() {
+                    return Ok(());
+                }
+
+                let threshold_signature = pks
+                    .combine_signatures(signature_shares.iter().map(|(peer, share)| (peer, share)))
+                    .expect("All signature shares are valid");
+
+                dbtx.remove_by_prefix(&ClientConfigSignatureSharePrefix)
+                    .await;
+
+                dbtx.insert_entry(
+                    &ClientConfigSignatureKey,
+                    &SerdeSignature(threshold_signature),
+                )
+                .await;
+
+                Ok(())
+            }
+        }
+    }
+
+    async fn request_signed_block(&self, index: u64) -> SignedBlock {
+        let keychain = self.keychain.clone();
+        let total_peers = self.keychain.peer_count();
+        let decoders = self.decoders();
+
+        let filter_map = move |response: SerdeModuleEncoding<SignedBlock>| match response
+            .try_into_inner(&decoders)
+        {
+            Ok(signed_block) => {
+                match signed_block.signatures.len() == keychain.threshold()
+                    && signed_block.signatures.iter().all(|(peer_id, sig)| {
+                        keychain.verify(
+                            &signed_block.block.header(index),
+                            sig,
+                            to_node_index(*peer_id),
+                        )
+                    }) {
+                    true => Ok(signed_block),
+                    false => Err(anyhow!("Invalid signatures")),
+                }
+            }
+            Err(error) => Err(anyhow!(error.to_string())),
+        };
+
+        let federation_api = WsFederationApi::new(self.api_endpoints.clone());
+
+        loop {
+            // we wait until we have stalled
+            sleep(Duration::from_secs(5)).await;
+
+            let result = federation_api
+                .request_with_strategy(
+                    FilterMap::new(filter_map.clone(), total_peers),
+                    AWAIT_SIGNED_BLOCK_ENDPOINT.to_string(),
+                    ApiRequestErased::new(index),
+                )
+                .await;
+
+            match result {
+                Ok(signed_block) => return signed_block,
+                Err(error) => tracing::error!("Error while requesting signed block: {}", error),
+            }
+        }
     }
 }
